requirements.txt
Django==1.7.3
Fabric==1.10.1
Markdown==2.5.2
PyYAML==3.11
Whoosh==2.4.0
amqp==1.4.6
amqplib==1.0.2
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.19
boto==2.36.0
celery==3.1.17
django-celery==3.1.16
django-endless-pagination==2.0
django-filter==0.9.1
django-haystack==2.0.0
django-referral==1.0.1
django-rest-swagger==0.2.8
django-storages==1.1.8
djangorestframework==3.0.3
docopt==0.4.0
ecdsa==0.13
fudge==1.0.3
kombu==3.0.24
mailchimp==2.0.9
paramiko==1.15.2
psycopg2==2.5.4
pycrypto==2.6.1
pytz==2014.10
requests==2.5.1
soundcloud==0.4.1
wsgiref==0.1.2
